+++ CHANGELOG.md
@@ -1,6 +1,13 @@
+## Unreleased
+
+- Add User Color to Message Header
+
 ## 0.1.0
 
 - **Added**
   - Minify `Simple Requests` module to minify the chat area it takes up
-  - Colorize `PF2e HUD`'s `Persistent HUD` to have color
-  - Add User Color to Message Header
+  - Colorize `PF2e HUD`'s `Persistent HUD`
+  - Colorize `PF2e Toolbelt`'s `Target Helper`
+  - Highlight `Pf2e Toolbelt`'s `Target Helper` unrolled rolls and make them glow
+  - Hide `Sell All Treasure`
+  - Hide `Default Craft Checks` (for Heroic Crafting)